+++ tests/e2e/specs/front-end/test-checkout-registered.js
@@ -1,0 +1,6 @@
+/*
+ * Internal dependencies
+ */
+const { runCheckoutRegisteredPageTest } = require( '@woocommerce/e2e-core-tests' );
+
+runCheckoutRegisteredPageTest();